kexec: __NR_kexec_file_load is set to undefined on LoongArch
The initial reason is that after the merger of 29fe5067ed07 ("kexec: make
-a the default"), kexec cannot be used on LoongArch architectures. We need
to add "-c" for normal use. The current kexec_file_load system call is not
implemented in architectures such as LoongArch, so it needs to pass kexec_load.
So we need to set __NR_kexec_file_load to undefined in unsupported architectures.
This will return EFALLBACK via is_kexec_file_load_implemented, and then via
kexec_load.

Signed-off-by: Youling Tang <tangyouling@loongson.cn>
Signed-off-by: Simon Horman <horms@kernel.org>