arm64: Scatter the logic of reading of kernel file into each probe
As more complicated capsule kernel format ,such as zboot, emerge, where
the compressed kernel is stored as a payload. The straight forward
decompression can not meet the demand.

Therefore, let a probe method read in the kernel file and decide how to
unfold the content by the method itself.

Since other arches still read the kernel image before probe method,
the purpose of the first argument differs from that in aarch64.

Signed-off-by: Pingfan Liu <piliu@redhat.com>
To: kexec@lists.infradead.org
Cc: horms@verge.net.au
Cc: ardb@kernel.org
Cc: jeremy.linton@arm.com