LoongArch: Change COMMAND_LINE_SIZE to 4096
The default COMMAND_LINE_SIZE inherited from asm-generic is 512, which
is too small for modern use cases. For example, when the parameter path
in the cmdline of the kdump first kernel is too long, the limit is
easily exceeded when loading the second kernel.

Increase the command line size to 4096 bytes, In LoongArch, this is
consistent with the upstream configuration of the Linux kernel as of
commit f7794a4d92ad ("LoongArch: Increase COMMAND_LINE_SIZE up to 4096").

This patch is necessary because the Linux kernel has been changed to
4096. When the kernel cmdline size exceeds 512, the --reuse-cmdline
option of the kexec command will inevitably result in an error.

The COMMAND_LINE_SIZE in kexec/arch/loongarch/crashdump-loongarch.h
is unused, so it is removed here.

Suggested-by: Youling Tang <tangyouling@kylinos.cn>
Signed-off-by: Chenghao Duan <duanchenghao@kylinos.cn>
Signed-off-by: Simon Horman <horms@kernel.org>